requirements.txt
fastapi
uvicorn
faiss-cpu
pillow
numpy
pydantic
markdownify
playwright
requests
beautifulsoup4
